+++ Elf.h
@@ -6,8 +6,8 @@ class Elf : public Player
    const int BASE_HP = 50;
    const int BASE_ATK = 25;
    const int BASE_DEF = 25;
-   const int SCORE_MODIFIER = 1;
-   const int GOLD_MODIFIER = 1;
+   const float SCORE_MODIFIER = 1;
+   const float GOLD_MODIFIER = 1;
 
 public:
    Elf() : Player{BASE_HP, BASE_ATK, BASE_DEF, SCORE_MODIFIER, GOLD_MODIFIER, "Elf"} {}
